Move `merge_nested_if`
