+++ crates/swc_ecma_minifier/src/compress/optimize/if_return.rs
@@ -1,4 +1,4 @@
-use swc_common::{util::take::Take, Spanned, DUMMY_SP};
+use swc_common::{util::take::Take, DUMMY_SP};
 use swc_ecma_ast::*;
 use swc_ecma_transforms_optimization::debug_assert_valid;
 use swc_ecma_utils::StmtLike;
@@ -12,36 +12,6 @@ use crate::{compress::util::is_pure_undefined, util::ExprOptExt};
 /// Methods related to the option `if_return`. All methods are noop if
 /// `if_return` is false.
 impl Optimizer<'_> {
-    pub(super) fn merge_nested_if(&mut self, s: &mut IfStmt) {
-        if !self.options.conditionals && !self.options.bools {
-            return;
-        }
-
-        if s.alt.is_some() {
-            return;
-        }
-
-        if let Stmt::If(IfStmt {
-            test,
-            cons,
-            alt: None,
-            ..
-        }) = &mut *s.cons
-        {
-            self.changed = true;
-            report_change!("if_return: Merging nested if statements");
-
-            s.test = BinExpr {
-                span: s.test.span(),
-                op: op!("&&"),
-                left: s.test.take(),
-                right: test.take(),
-            }
-            .into();
-            s.cons = cons.take();
-        }
-    }
-
     pub(super) fn merge_if_returns(
         &mut self,
         stmts: &mut Vec<Stmt>,
